feat: modify pause & abort task
