+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eko-ai/eko",
-  "version": "2.1.7-alpha.3",
+  "version": "2.1.7-alpha.4",
   "description": "Empowering language to transform human words into action.",
   "workspaces": [
     "packages/eko-core",
