+++ packages/eko-core/src/core/context.ts
@@ -1,7 +1,12 @@
 import { Agent } from "../agent";
 import { sleep } from "../common/utils";
 import Chain, { AgentChain } from "./chain";
-import { EkoConfig, Workflow, WorkflowAgent } from "../types";
+import {
+  EkoConfig,
+  LanguageModelV1Prompt,
+  Workflow,
+  WorkflowAgent,
+} from "../types";
 
 export default class Context {
   taskId: string;
@@ -11,8 +16,9 @@ export default class Context {
   controller: AbortController;
   variables: Map<string, any>;
   workflow?: Workflow;
-  paused: boolean = false;
   conversation: string[] = [];
+  private pauseStatus: 0 | 1 | 2 = 0;
+  readonly currentStepControllers: Set<AbortController> = new Set();
 
   constructor(
     taskId: string,
@@ -29,14 +35,19 @@ export default class Context {
   }
 
   async checkAborted() {
-    // this.controller.signal.throwIfAborted();
     if (this.controller.signal.aborted) {
       const error = new Error("Operation was interrupted");
       error.name = "AbortError";
       throw error;
     }
-    while (this.paused) {
+    while (this.pauseStatus > 0) {
       await sleep(500);
+      if (this.pauseStatus == 2) {
+        this.currentStepControllers.forEach((c) => {
+          c.abort("Pause");
+        });
+        this.currentStepControllers.clear();
+      }
       if (this.controller.signal.aborted) {
         const error = new Error("Operation was interrupted");
         error.name = "AbortError";
@@ -59,6 +70,20 @@ export default class Context {
     const agentContext = agent["agentContext"] as AgentContext;
     return [agent, agentNode.agent, agentContext];
   }
+
+  get pause() {
+    return this.pauseStatus > 0;
+  }
+
+  setPause(pause: boolean, abortCurrentStep?: boolean) {
+    this.pauseStatus = pause ? (abortCurrentStep ? 2 : 1) : 0;
+    if (this.pauseStatus == 2) {
+      this.currentStepControllers.forEach((c) => {
+        c.abort("Pause");
+      });
+      this.currentStepControllers.clear();
+    }
+  }
 }
 
 export class AgentContext {
@@ -67,6 +92,7 @@ export class AgentContext {
   agentChain: AgentChain;
   variables: Map<string, any>;
   consecutiveErrorNum: number;
+  messages?: LanguageModelV1Prompt;
 
   constructor(context: Context, agent: Agent, agentChain: AgentChain) {
     this.context = context;
