+++ packages/eko-nodejs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eko-ai/eko-nodejs",
-  "version": "2.1.7-alpha.3",
+  "version": "2.1.7-alpha.4",
   "description": "Empowering language to transform human words into action.",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
